requirements.txt
streamlit
groq
python-dotenv
faiss-cpu
pypdf
jina
pillow